sp: re-structure unit test code
