+++ include/fluent-bit/stream_processor/flb_sp.h
@@ -35,6 +35,11 @@
 #define FLB_SP_BOOLEAN       2
 #define FLB_SP_STRING        3
 
+struct sp_buffer {
+    char* buffer;
+    size_t size;
+};
+
 struct aggregate_num {
     int type;
     int ops;
@@ -159,12 +164,20 @@ void flb_sp_destroy(struct flb_sp *sp);
 int flb_sp_do(struct flb_sp *sp, struct flb_input_instance *in,
               const char *tag, int tag_len,
               const char *buf_data, size_t buf_size);
-int flb_sp_test_do(struct flb_sp *sp, struct flb_sp_task *task,
-                   const char *tag, int tag_len,
-                   const char *buf_data, size_t buf_size,
-                   char **out_data, size_t *out_size);
-int flb_sp_test_fd_event(int fd, struct flb_sp_task *task, char **out_data,
-                         size_t *out_size);
+int sp_process_data(const char *tag, int tag_len,
+                    const char *buf_data, size_t buf_size,
+                    char **out_buf, size_t *out_size,
+                    struct flb_sp_task *task,
+                    struct flb_sp *sp);
+int sp_process_data_aggr(const char *buf_data, size_t buf_size,
+                         const char *tag, int tag_len,
+                         struct flb_sp_task *task,
+                         struct flb_sp *sp);
+void package_results(const char *tag, int tag_len,
+                     char **out_buf, size_t *out_size,
+                     struct flb_sp_task *task);
+int sp_process_hopping_slot(const char *tag, int tag_len,
+                            struct flb_sp_task *task);
 
 int flb_sp_snapshot_create(struct flb_sp_task *task);
 struct flb_sp_task *flb_sp_task_create(struct flb_sp *sp, const char *name,
